Add usemin-handler task, moved getBlocks as usemin:blocks helper
Based off @necolas' idea and initial work:

- https://gist.github.com/3024891

The goal of the usemin task is to consider your HTML markup as the primary
source of information. All the path and reference to assets to optimize should
be there, and all the necessary info for setting up the build script should be
available.

Usemin-handler task doesn't do any actual optimizations, it's job is to init
the Grunt configuration automatically from HTML markup.